# Week09/Dawon-Y/mission/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiResponse__memberMissionId-number__": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"properties": {
							"memberMissionId": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"memberMissionId"
						],
						"type": "object"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Awaited_ReturnType_typeoflistOngoingMissions__": {
				"properties": {
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					},
					"data": {
						"items": {
							"properties": {
								"mission": {
									"properties": {
										"missionSpec": {
											"type": "string"
										},
										"deadline": {
											"type": "string",
											"format": "date-time"
										},
										"reward": {
											"type": "number",
											"format": "double"
										},
										"store": {
											"properties": {
												"name": {
													"type": "string"
												}
											},
											"required": [
												"name"
											],
											"type": "object"
										}
									},
									"required": [
										"missionSpec",
										"deadline",
										"reward",
										"store"
									],
									"type": "object"
								},
								"status": {
									"type": "string"
								},
								"id": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"mission",
								"status",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"pagination",
					"data"
				],
				"type": "object",
				"description": "Recursively unwraps the \"awaited type\" of a type. Non-promise \"thenables\" should resolve to `never`. This emulates the behavior of `await`."
			},
			"ApiResponse_Awaited_ReturnType_typeoflistOngoingMissions___": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/Awaited_ReturnType_typeoflistOngoingMissions__"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Awaited_ReturnType_typeofupdateMissionComplete__": {
				"properties": {
					"missionId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"memberId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"missionId",
					"status",
					"updatedAt",
					"createdAt",
					"memberId",
					"id"
				],
				"type": "object",
				"description": "Recursively unwraps the \"awaited type\" of a type. Non-promise \"thenables\" should resolve to `never`. This emulates the behavior of `await`."
			},
			"ApiResponse_Awaited_ReturnType_typeofupdateMissionComplete___": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/Awaited_ReturnType_typeofupdateMissionComplete__"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"paths": {
...
		"/missions/{missionId}/challenge": {
			"post": {
				"operationId": "HandleChallengeMission",
				"responses": {
					"201": {
						"description": "미션 도전 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__memberMissionId-number__"
								}
							}
						}
					},
					"401": {
						"description": "인증 실패"
					},
					"409": {
						"description": "이미 도전 중인 미션 (M001)"
					}
				},
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/members/{memberId}/missions": {
			"get": {
				"operationId": "HandleListOngoingMissions",
				"responses": {
					"200": {
						"description": "미션 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Awaited_ReturnType_typeoflistOngoingMissions___"
								}
							}
						}
					},
					"401": {
						"description": "인증 실패"
					}
				},
				"tags": [
					"Members"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "memberId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/member-missions/{memberMissionId}/complete": {
			"patch": {
				"operationId": "HandleCompleteMission",
				"responses": {
					"200": {
						"description": "미션 완료 처리 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Awaited_ReturnType_typeofupdateMissionComplete___"
								}
							}
						}
					},
					"401": {
						"description": "인증 실패"
					}
				},
				"tags": [
					"MemberMissions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "memberMissionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}